Support Go 1.6 (#120)
* all: use golang.org/x/net/context

* internal: use ctxhttp to use /x/ context

The 1.6 Request type doesn't have WithContext.

* all: don't use subtests to keep 1.6 compatibility

* integration: use float64 for fields with exp value

Values like -7e+07 cannot be parsed into ints in Go 1.6. So, use
floats instead.

* integration/messaging: use t.Fatal not log.Fatal

* travis: add 1.6.x

* changelog: mention addition of 1.6 support

* readme: mention go version support

+++ integration/db/db_test.go
@@ -690,11 +690,11 @@ func protectedRef(t *testing.T, p string) string {
 }
 
 type Dinosaur struct {
-	Appeared int     `json:"appeared"`
+	Appeared float64 `json:"appeared"`
 	Height   float64 `json:"height"`
 	Length   float64 `json:"length"`
 	Order    string  `json:"order"`
-	Vanished int     `json:"vanished"`
+	Vanished float64 `json:"vanished"`
 	Weight   int     `json:"weight"`
 	Ratings  Ratings `json:"ratings"`
 }
